+++ src/formbuilder/formbuilder-application/src/views/Form.tsx
@@ -5,12 +5,14 @@ import Section from '../components/Section';
 import {
     FormContext,
     addNewSection,
+    duplicateSection,
     removeSection,
     swapSection,
     swapQuestion,
 } from '../store/FormStore';
 import * as DND from 'react-beautiful-dnd';
 import './Form.css';
+import ISection from '../types/ISection';
 
 function CreateForm(): JSX.Element {
     const [i, setI] = useState(0);
@@ -18,14 +20,19 @@ function CreateForm(): JSX.Element {
 
     const [dragIndex, setDragIndex] = useState(-1);
 
+    const [collapsedSection, setCollapsedSection] = useState('A');
+
     function dispatchAddNewSection(index?: number) {
-        setI(i + 1);
-        //const newSectionAction = addNewSection();
         dispatch(addNewSection());
     }
 
     function dispatchRemoveSection(index: number) {
-        dispatch(removeSection(index));
+        if (window.confirm('Vil du slette denne seksjonen?'))
+            dispatch(removeSection(index));
+    }
+
+    function dispatchDuplicateSection(index: number, id: string) {
+        dispatch(duplicateSection(index, id));
     }
 
     function onDragEnd(result: DND.DropResult) {
@@ -36,6 +43,7 @@ function CreateForm(): JSX.Element {
 
         if (result.type === 'section') {
             dispatch(swapSection(sourceIndex, destIndex));
+            setCollapsedSection('A');
         } else if (result.type === 'question') {
             const sourceParentId = result.source.droppableId;
             const destParentId = result.destination.droppableId;
@@ -50,10 +58,35 @@ function CreateForm(): JSX.Element {
         }
     }
 
-    function onDragStart(startResponder: DND.DragStart) {
-        if (startResponder.type === 'section')
-            setDragIndex(startResponder.source.index);
+    function onBeforeCapture(startResponder: DND.BeforeCapture) {
+        setCollapsedSection(startResponder.draggableId);
     }
+
+    type memoSection = {
+        sectionId: string;
+        index: number;
+        provided: DND.DraggableProvided;
+    };
+
+    function compareSections(
+        prevSection: memoSection,
+        nextSection: memoSection,
+    ) {
+        return prevSection.index === nextSection.index;
+    }
+
+    // const SectionMemo = React.memo((props: memoSection) => {
+    //     return (
+    //         <Section
+    //             key={props.sectionId}
+    //             sectionId={props.sectionId}
+    //             removeSection={() => dispatchRemoveSection(props.index)}
+    //             provided={props.provided}
+    //             collapsed={props.sectionId === collapsedSection}
+    //             index={props.index}
+    //         />
+    //     );
+    // }, compareSections);
 
     return (
         <div>
@@ -65,7 +98,7 @@ function CreateForm(): JSX.Element {
             <Row style={{ margin: '61px 0 0 0' }}>
                 <DND.DragDropContext
                     onDragEnd={onDragEnd}
-                    onBeforeDragStart={onDragStart}
+                    onBeforeCapture={onBeforeCapture}
                 >
                     <DND.Droppable droppableId="section" type="section">
                         {(provided, snapshot) => (
@@ -79,12 +112,10 @@ function CreateForm(): JSX.Element {
                                 >
                                     {state.sectionOrder.map(
                                         (sectionId: string, index: number) => {
-                                            const section =
-                                                state.sections[sectionId];
                                             return (
                                                 <DND.Draggable
-                                                    key={section.id}
-                                                    draggableId={section.id}
+                                                    key={'drag' + sectionId}
+                                                    draggableId={sectionId}
                                                     index={index}
                                                 >
                                                     {(provided, snapshot) => (
@@ -94,10 +125,26 @@ function CreateForm(): JSX.Element {
                                                             }
                                                             {...provided.draggableProps}
                                                         >
+                                                            {/* <SectionMemo
+                                                                key={sectionId}
+                                                                sectionId={
+                                                                    sectionId
+                                                                }
+                                                                index={index}
+                                                                provided={
+                                                                    provided
+                                                                }
+                                                            /> */}
                                                             <Section
-                                                                key={section.id}
-                                                                section={
-                                                                    section
+                                                                key={sectionId}
+                                                                sectionId={
+                                                                    sectionId
+                                                                }
+                                                                duplicateSection={() =>
+                                                                    dispatchDuplicateSection(
+                                                                        index,
+                                                                        sectionId,
+                                                                    )
                                                                 }
                                                                 removeSection={() =>
                                                                     dispatchRemoveSection(
@@ -108,8 +155,8 @@ function CreateForm(): JSX.Element {
                                                                     provided
                                                                 }
                                                                 collapsed={
-                                                                    index ===
-                                                                    dragIndex
+                                                                    sectionId ===
+                                                                    collapsedSection
                                                                 }
                                                                 index={index}
                                                             />
